refactored alert box and changed percentage to int

+++ quiz_app - Refactored/lib/ui/Screens/Question/models/scores.dart
@@ -20,7 +20,7 @@ class CourseScore {
   String? courseType;
   int? courseScore;
   int? userId;
-  double? coursePercentage;
+  int? coursePercentage;
 
   factory CourseScore.fromJson(Map<String, dynamic> json) => CourseScore(
         courseId: json["courseId"],
